+++ apps/admin/src/app/api/auth/[...nextauth]/route.ts
@@ -1,137 +1,6 @@
-import NextAuth, { NextAuthOptions } from 'next-auth';
-import CredentialsProvider from 'next-auth/providers/credentials';
-import { login } from '@/server/domains/account';
-import { LoginUser } from '@/features/auth';
+import NextAuth from 'next-auth';
+import { authOptions } from './options';
 require('@/libs/prisma');
-
-// For more information on each option (and a full list of options) go to
-// https://next-auth.js.org/configuration/options
-const authOptions: NextAuthOptions = {
-  // https://next-auth.js.org/configuration/providers
-  providers: [
-    CredentialsProvider({
-      // The name to display on the sign in form (e.g. 'Sign in with...')
-      name: 'Credentials',
-      // The credentials is used to generate a suitable form on the sign in page.
-      // You can specify whatever fields you are expecting to be submitted.
-      // e.g. domain, username, password, 2FA token, etc.
-      credentials: {
-        username: { label: 'email', type: 'text', placeholder: 'sample@sample.com' },
-        password: { label: 'password', type: 'password' },
-      },
-      async authorize(credentials, req) {
-        // Add logic here to look up the user from the credentials supplied
-        // const connection = await createConnection('default');
-        try {
-          if (credentials) {
-            const loginUser = await login(credentials.username, credentials.password);
-            if (loginUser) {
-              // Any object returned will be saved in `user` property of the JWT
-              const user: LoginUser = {
-                id: loginUser.id,
-                email: loginUser.email,
-                status: loginUser.status,
-                privilege: loginUser.privilege,
-                caution: loginUser.caution,
-                name: loginUser.name,
-                image: loginUser.image,
-              };
-              return user;
-            } else {
-              // If you return null or false then the credentials will be rejected
-              return null;
-              // You can also Reject this callback with an Error or with a URL:
-              // throw new Error('login failed'); // Redirect to error page
-              // throw '/path/to/redirect'        // Redirect to a URL
-            }
-          }
-        } catch (e) {
-          console.log(e);
-        }
-
-        return null;
-      },
-    }),
-  ],
-
-  // The secret should be set to a reasonably long random string.
-  // It is used to sign cookies and to sign and encrypt JSON Web Tokens, unless
-  // a separate secret is defined explicitly for encrypting the JWT.
-  secret: process.env.NEXTAUTH_SECRET,
-
-  session: {
-    // Use JSON Web Tokens for session instead of database sessions.
-    // This option can be used with or without a database for users/accounts.
-    // Note: `strategy` should be set to 'jwt' if no database is used.
-    strategy: 'jwt',
-
-    // Seconds - How long until an idle session expires and is no longer valid.
-    maxAge: 1 * 24 * 60 * 60, // 1 days
-
-    // Seconds - Throttle how frequently to write to database to extend a session.
-    // Use it to limit write operations. Set to 0 to always update the database.
-    // Note: This option is ignored if using JSON Web Tokens
-    // updateAge: 24 * 60 * 60, // 24 hours
-  },
-
-  // JSON Web tokens are only used for sessions if the `jwt: true` session
-  // option is set - or by default if no database is specified.
-  // https://next-auth.js.org/configuration/options#jwt
-  jwt: {
-    // You can define your own encode/decode functions for signing and encryption
-    // if you want to override the default behaviour.
-    // encode: async ({ secret, token, maxAge }) => {},
-    // decode: async ({ secret, token, maxAge }) => {},
-  },
-
-  // You can define custom pages to override the built-in ones. These will be regular Next.js pages
-  // so ensure that they are placed outside of the '/api' folder, e.g. signIn: '/auth/mycustom-signin'
-  // The routes shown here are the default URLs that will be used when a custom
-  // pages is not specified for that route.
-  // https://next-auth.js.org/configuration/pages
-  pages: {
-    signIn: '/auth/signin', // Displays signin buttons
-    signOut: '/', // Displays form with sign out button
-    error: '/auth/error', // Error code passed in query string as ?error=
-    // verifyRequest: '/auth/verify-request', // Used for check email page
-    // newUser: null // If set, new users will be directed here on first sign in
-  },
-
-  // Callbacks are asynchronous functions you can use to control what happens
-  // when an action is performed.
-  // https://next-auth.js.org/configuration/callbacks
-  callbacks: {
-    async signIn({ user, account, profile, email, credentials }) {
-      return true;
-    },
-    // async redirect({ url, baseUrl }) { return baseUrl },
-    async session({ session, token, user }) {
-      session.user.status = (token.user as LoginUser).status;
-      session.user.privilege = (token.user as LoginUser).privilege;
-      session.user.caution = (token.user as LoginUser).caution;
-      return session;
-    },
-    async jwt({ token, user, account, profile }) {
-      if (user) {
-        token.user = user;
-      }
-      return token;
-    },
-  },
-
-  // Events are useful for logging
-  // https://next-auth.js.org/configuration/events
-  events: {},
-
-  // You can set the theme to 'light', 'dark' or use 'auto' to default to the
-  // whatever prefers-color-scheme is set to in the browser. Default is 'auto'
-  theme: {
-    colorScheme: 'auto',
-  },
-
-  // Enable debug messages in the console if you are having problems
-  debug: process.env.NODE_ENV !== 'production',
-};
 
 const handler = NextAuth(authOptions);
