+++ apps/admin/src/layouts/auth/layout.stories.tsx
@@ -1,4 +1,3 @@
-
 import { SignIn } from '@/features/auth/components/signin';
 import { Layout } from './index';
 import type { Meta, StoryFn } from '@storybook/react';
@@ -13,7 +12,6 @@ const Template: StoryFn<typeof Layout> = (args) => <Layout {...args} />;
 export const Sample1 = Template.bind({});
 Sample1.storyName = 'default';
 Sample1.args = {
-  title: 'Page title',
   children: <SignIn />,
 };
 Sample1.parameters = {
